Pre-fill DateTime caches when constructing from civil components.
The basic idea here, is that if we already know the civil and
time components when the DateTime is created, we can pre-fill
Date's caches, so that the (slow) functions DateTime.jd_to_civil
etc are never called.

This patch relies on knowledge of how Date's caching works (see
the Date.once function), ie how the instance variable names are
constructed. It won't cause any harm though if its changed though.