Reimplement DBI::Row#dup instead of aliasing to clone
Though dup has been aliased to clone for ages, at some point clone
was changed to actually be a deep clone via Marshal. As every Row
in a #select_all is dup'd, this adds a large overhead, both in time
(especially for more types implemented in ruby like DateTime) and
also in memory (every row has a *different* copy of the column names
and types etc).

This restores the old behaviour, which means that Row#dup provides
a row which shares the column meta data, but dups the array. The
implementation is a bit messy (listing instance vars) - but super
just calls Array#dup. Other option might be to re-bind Object#dup.

Note that this change is also needed for a future patch which
implements lazy DateTime ajd calclation (which marshalling forces
unnecessarily).